refactor(utils): moved types to types folder
