+++ src/api/discord.ts
@@ -1,7 +1,7 @@
 import { signOut } from 'next-auth/react';
 import { callReturn } from '@/utils/fetch/core';
 import { discordRequest } from '@/utils/fetch/requests';
-import type { Guild, UserInfo } from '@/utils/types';
+import type { Guild, UserInfo } from '@/types/types';
 
 export async function fetchUserInfo(accessToken: string) {
 	return await callReturn<UserInfo>(
