+++ src/components/forms/ChannelSelect.tsx
@@ -5,10 +5,9 @@ import { common } from '@/config/translations/common';
 import { useGuildChannelsQuery } from '@/api/hooks';
 import { MdRecordVoiceOver } from 'react-icons/md';
 import { useController } from 'react-hook-form';
-import { ChannelTypes, GuildChannel } from '@/utils/types';
+import { ChannelTypes, GuildChannel, Override } from '@/types/types';
 import { forwardRef, useMemo } from 'react';
 import { ControlledInput } from './types';
-import { Override } from '@/utils/types';
 import { Icon } from '@chakra-ui/react';
 import { useRouter } from 'next/router';
 import { FormCard } from './Form';
